S02P13C110-66 [+] 광고주 메인페이지에 차트 추가
