+++ MANAGE_FRONTEND/package.json
@@ -19,6 +19,7 @@
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.3.0",
     "react-swipeable-views": "^0.13.7",
+    "recharts": "^2.0.0-beta.1",
     "redux": "^4.0.5",
     "redux-actions": "^2.6.5",
     "redux-devtools-extension": "^2.13.8",
