+++ MANAGE_FRONTEND/src/App.js
@@ -3,7 +3,6 @@ import { Route } from 'react-router-dom';
 import MainPage from './pages/MainPage';
 import LoginPage from './pages/LoginPage';
 import RegisterPage from './pages/RegisterPage';
-import ManagePage from './pages/ManagePage';
 import AdminPage from './pages/AdminPage';
 import DetailPage from './pages/DetailPage';
 
@@ -13,7 +12,6 @@ const App = () => {
       <Route component={MainPage} path="/" exact />
       <Route component={LoginPage} path="/login" />
       <Route component={RegisterPage} path="/register" />
-      <Route component={ManagePage} path="/manage" />
       <Route component={AdminPage} path="/admin" />
       <Route component={DetailPage} path="/detail/:id" />
     </>
