--- MANAGE_FRONTEND/src/components/manage/ContentContainer.js
@@ -1,65 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import SwipeableViews from 'react-swipeable-views';
-import Typography from '@material-ui/core/Typography';
-import Box from '@material-ui/core/Box';
-import { makeStyles, useTheme } from '@material-ui/core/styles';
-
-function TabPanel(props) {
-  const { children, value, index, ...other } = props;
-
-  return (
-    <Typography
-      component="div"
-      role="tabpanel"
-      hidden={value !== index}
-      id={`full-width-tabpanel-${index}`}
-      aria-labelledby={`full-width-tab-${index}`}
-      {...other}
-    >
-      {value === index && <Box p={3}>{children}</Box>}
-    </Typography>
-  );
-}
-
-TabPanel.propTypes = {
-  children: PropTypes.node,
-  index: PropTypes.any.isRequired,
-  value: PropTypes.any.isRequired,
-};
-
-const useStyles = makeStyles(theme => ({
-  content: {
-    backgroundColor: theme.palette.background.paper,
-    width: '100%',
-    textAlign: 'center',
-  },
-}));
-
-const ContentContainer = ({ value }) => {
-  const classes = useStyles();
-  const theme = useTheme();
-  return (
-    <SwipeableViews
-      className={classes.content}
-      axis={theme.direction === 'rtl' ? 'x-reverse' : 'x'}
-      index={value}
-      // onChangeIndex={handleChangeIndex}
-    >
-      <TabPanel value={value} index={0} dir={theme.direction}>
-        Item 1
-      </TabPanel>
-      <TabPanel value={value} index={1} dir={theme.direction}>
-        Item 2
-      </TabPanel>
-      <TabPanel value={value} index={2} dir={theme.direction}>
-        Item 3
-      </TabPanel>
-      <TabPanel value={value} index={3} dir={theme.direction}>
-        Item 4
-      </TabPanel>
-    </SwipeableViews>
-  );
-};
-
-export default ContentContainer;
