--- MANAGE_FRONTEND/src/components/manage/TabContainer.js
@@ -1,54 +1,0 @@
-import React, { useState } from 'react';
-
-import { makeStyles, useTheme } from '@material-ui/core/styles';
-import AppBar from '@material-ui/core/AppBar';
-import Tabs from '@material-ui/core/Tabs';
-import Tab from '@material-ui/core/Tab';
-
-function a11yProps(index) {
-  return {
-    id: `full-width-tab-${index}`,
-    'aria-controls': `full-width-tabpanel-${index}`,
-  };
-}
-
-const useStyles = makeStyles(theme => ({
-  root: {
-    backgroundColor: theme.palette.background.paper,
-    width: '100%',
-  },
-}));
-
-const TabContainer = ({ value, setValue }) => {
-  const classes = useStyles();
-
-  const handleChange = (event, newValue) => {
-    setValue(newValue);
-  };
-
-  const handleChangeIndex = index => {
-    setValue(index);
-  };
-
-  return (
-    <div className={classes.root}>
-      <AppBar position="static" color="default">
-        <Tabs
-          value={value}
-          onChange={handleChange}
-          indicatorColor="primary"
-          textColor="primary"
-          variant="fullWidth"
-          aria-label="full width tabs example"
-        >
-          <Tab label="Item 1" {...a11yProps(0)} />
-          <Tab label="Item 2" {...a11yProps(1)} />
-          <Tab label="Item 3" {...a11yProps(2)} />
-          <Tab label="Item 4" {...a11yProps(3)} />
-        </Tabs>
-      </AppBar>
-    </div>
-  );
-};
-
-export default TabContainer;
